minor fixes 2.0
