+++ kafka-connect-upload-source/src/main/java/org/radarbase/connect/upload/converter/altoida/AltoidaAccelerometerConverter.kt
@@ -1,5 +1,7 @@
-package org.radarbase.connect.upload.converter
+package org.radarbase.connect.upload.converter.altoida
 
+import org.radarbase.connect.upload.converter.CsvRecordConverter
+import org.radarbase.connect.upload.converter.TopicData
 import org.radarcns.connector.altoida.AltoidaAcceleration
 
 class AltoidaAccelerometerConverter(override val sourceType: String = "acceleration", val topic: String = "connect_upload_altoida_acceleration")
