+++ kafka-connect-upload-source/src/main/java/org/radarbase/connect/upload/converter/altoida/AltoidaActionConverter.kt
@@ -1,5 +1,7 @@
-package org.radarbase.connect.upload.converter
+package org.radarbase.connect.upload.converter.altoida
 
+import org.radarbase.connect.upload.converter.CsvRecordConverter
+import org.radarbase.connect.upload.converter.TopicData
 import org.radarcns.connector.altoida.AltoidaAction
 
 class AltoidaActionConverter(override val sourceType: String = "log", val topic: String = "connect_upload_altoida_action")
