+++ kafka-connect-upload-source/src/main/java/org/radarbase/connect/upload/converter/altoida/AltoidaObjectsConverter.kt
@@ -1,5 +1,7 @@
-package org.radarbase.connect.upload.converter
+package org.radarbase.connect.upload.converter.altoida
 
+import org.radarbase.connect.upload.converter.CsvRecordConverter
+import org.radarbase.connect.upload.converter.TopicData
 import org.radarcns.connector.altoida.AltoidaObjects
 
 class AltoidaObjectsConverter(override val sourceType: String = "objects", val topic: String = "connect_upload_altoida_objects")
@@ -17,7 +19,7 @@ class AltoidaObjectsConverter(override val sourceType: String = "objects", val t
                 lineValues["X"]?.toFloat(),
                 lineValues["Y"]?.toFloat(),
                 lineValues["Z"]?.toFloat(),
-        )
+                )
 
         return TopicData(false, topic, objects)
     }
